9479: rename ApplyStamp folder to StampMotion
