+++ web-client/src/presenter/actions/StampMotion/clearOptionalFieldsStampFormAction.js
