+++ web-client/src/presenter/actions/StampMotion/clearPDFStampDataAction.js
