+++ web-client/src/presenter/actions/StampMotion/clearPDFStampDataAction.test.js
