+++ web-client/src/presenter/actions/StampMotion/unsetDeniedOptionsOnStampFormAction.js
