+++ web-client/src/presenter/actions/StampMotion/unsetDeniedOptionsOnStampFormAction.test.js
