+++ web-client/src/presenter/actions/StampMotion/validateStampAction.js
