+++ web-client/src/presenter/actions/StampMotion/validateStampAction.test.js
