+++ web-client/src/presenter/sequences/clearDueDateSequence.js
@@ -1,3 +1,3 @@
-import { clearDueDateAction } from '../actions/ApplyStamp/clearDueDateAction';
+import { clearDueDateAction } from '../actions/StampMotion/clearDueDateAction';
 
 export const clearDueDateSequence = [clearDueDateAction];
