+++ web-client/src/presenter/sequences/clearOptionalFieldsStampFormSequence.js
@@ -1,6 +1,6 @@
 import { clearAlertsAction } from '../actions/clearAlertsAction';
 import { clearErrorAlertsAction } from '../actions/clearErrorAlertsAction';
-import { clearOptionalFieldsStampFormAction } from '../actions/ApplyStamp/clearOptionalFieldsStampFormAction';
+import { clearOptionalFieldsStampFormAction } from '../actions/StampMotion/clearOptionalFieldsStampFormAction';
 import { stopShowValidationAction } from '../actions/stopShowValidationAction';
 
 export const clearOptionalFieldsStampFormSequence = [
