+++ web-client/src/views/StampMotion/ApplyStamp.jsx
